Attempt to unmess bitwidth/bitcount types (it's always u32)

+++ src/bitstream/bitstream/decoder/bitstreamer/jpeg/tests.rs
@@ -4,7 +4,7 @@ use super::*;
 
 #[test]
 fn byte_enumeration_test() -> Result<(), Box<dyn core::error::Error>> {
-    const NUM_BITS: usize = 8;
+    const NUM_BITS: u32 = 8;
     let inputs: Vec<Vec<u8>> = vec![
         vec![1, 0, 0, 0],
         vec![1, 2, 0, 0],
@@ -35,7 +35,7 @@ fn byte_enumeration_test() -> Result<(), Box<dyn core::error::Error>> {
 
 #[test]
 fn nibble_enumeration_test() -> Result<(), Box<dyn core::error::Error>> {
-    const NUM_BITS: usize = 4;
+    const NUM_BITS: u32 = 4;
     let inputs: Vec<Vec<u8>> = vec![
         vec![16, 0, 0, 0],
         vec![18, 0, 0, 0],
@@ -73,7 +73,7 @@ fn nibble_enumeration_test() -> Result<(), Box<dyn core::error::Error>> {
 
 #[test]
 fn bit_enumeration_test() -> Result<(), Box<dyn core::error::Error>> {
-    const NUM_BITS: usize = 1;
+    const NUM_BITS: u32 = 1;
     let inputs: Vec<Vec<u8>> = vec![
         vec![128, 0, 0, 0],
         vec![64, 0, 0, 0],
@@ -128,7 +128,7 @@ fn bit_enumeration_test() -> Result<(), Box<dyn core::error::Error>> {
 
 #[test]
 fn sliding_0xff_test() -> Result<(), Box<dyn core::error::Error>> {
-    const NUM_BITS: usize = 1;
+    const NUM_BITS: u32 = 1;
     let inputs: Vec<Vec<u8>> = vec![
         vec![0xFF, 0x00, 0, 0, 0],
         vec![127, 128, 0, 0],
@@ -177,7 +177,7 @@ fn sliding_0xff_test() -> Result<(), Box<dyn core::error::Error>> {
 #[test]
 fn sliding_0xff_prefixed_by_enumerated_nibbles_test()
 -> Result<(), Box<dyn core::error::Error>> {
-    const NUM_BITS: usize = 4;
+    const NUM_BITS: u32 = 4;
     let inputs: Vec<Vec<u8>> = vec![
         vec![0xFF, 0x00, 0, 0, 0],
         vec![31, 240, 0, 0],
@@ -220,7 +220,7 @@ fn sliding_0xff_prefixed_by_enumerated_nibbles_test()
 #[test]
 fn sliding_0xff_through_enumerated_bytes_test()
 -> Result<(), Box<dyn core::error::Error>> {
-    const NUM_BITS: usize = 8;
+    const NUM_BITS: u32 = 8;
     let inputs: Vec<Vec<u8>> = vec![
         vec![
             0xFF, 0x00, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15,
@@ -303,7 +303,7 @@ fn sliding_0xff_through_enumerated_bytes_test()
 #[test]
 fn sliding_0xff_through_enumerated_nibbles_test()
 -> Result<(), Box<dyn core::error::Error>> {
-    const NUM_BITS: usize = 4;
+    const NUM_BITS: u32 = 4;
     let inputs: Vec<Vec<u8>> = vec![
         vec![0xFF, 0x00, 18, 52, 86, 120, 154, 188, 222],
         vec![31, 242, 52, 86, 120, 154, 188, 222],
@@ -353,7 +353,7 @@ fn sliding_0xff_through_enumerated_nibbles_test()
 #[test]
 fn sliding_0xff_non_0x00_control_sequence_test()
 -> Result<(), Box<dyn core::error::Error>> {
-    const NUM_BITS: usize = 8;
+    const NUM_BITS: u32 = 8;
     let inputs: Vec<Vec<u8>> = vec![
         vec![
             0xFF, 0x01, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15,
